+++ mobile/android/android-components/components/feature/toolbar/src/main/java/mozilla/components/feature/toolbar/ToolbarInteractor.kt
@@ -30,13 +30,7 @@ class ToolbarInteractor(
             } else {
                 searchUseCase?.invoke(text) ?: loadUrlUseCase.invoke(text)
             }
+            true
         }
-    }
-
-    /**
-     * Stops this interactor.
-     */
-    fun stop() {
-        toolbar.setOnUrlCommitListener { }
     }
 }