+++ mobile/android/android-components/components/feature/toolbar/src/test/java/mozilla/components/feature/toolbar/ToolbarAutocompleteFeatureTest.kt
@@ -49,7 +49,7 @@ class ToolbarAutocompleteFeatureTest {
             return false
         }
 
-        override fun setOnUrlCommitListener(listener: (String) -> Unit) {
+        override fun setOnUrlCommitListener(listener: (String) -> Boolean) {
             fail()
         }