+++ mobile/android/android-components/samples/toolbar/src/main/java/org/mozilla/samples/toolbar/ToolbarActivity.kt
@@ -406,6 +406,8 @@ class ToolbarActivity : AppCompatActivity() {
             if (url.isNotEmpty()) {
                 toolbar.url = url
             }
+
+            true
         }
     }